requirements.txt
anyio==3.5.0
async-generator==1.10
attrs==21.4.0
Authlib==1.0.1
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
cycler==0.11.0
DateTime==4.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
fonttools==4.31.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
kiwisolver==1.4.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
multitasking==0.0.10
mysql-connector==2.2.9
numpy==1.22.3
opstrat==0.1.7
outcome==1.1.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
prompt-toolkit==3.0.29
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rfc3986==1.5.0
scipy==1.8.0
seaborn==0.11.2
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
tda-api==1.5.3
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.0.3
wsproto==1.1.0
yfinance==0.1.70
zope.interface==5.4.0
